requirements.txt
annotated-types==0.6.0
anyio==4.3.0
atlastk==0.13.3
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
fastapi==0.110.3
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
itsdangerous==2.2.0
passlib==1.7.4
psycopg2==2.9.9
pydantic-core==2.18.2
pydantic-settings==2.2.1
pydantic==2.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
setuptools==60.2.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tortoise==0.1.1
typing-extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wheel==0.37.1
